+++ software/appver.pas
@@ -15,7 +15,7 @@ unit appver;
 interface
 
 const
-  PROX_VERSION = '4.34.0.0';
+  PROX_VERSION = '4.35.0.0';
 
 implementation